carProjectOfUserInterface/x86/requirements.txt
matplotlib
librosa==0.8.1
numpy==1.22.0
pygame==2.5.0
pyserial==3.5
ttkbootstrap==1.10.1
